Update - Start to merging the cell detection part (model_2)
