--- app/requirements_model_1.txt
@@ -1,152 +1,0 @@
-aiofiles==22.1.0
-aiosqlite==0.19.0
-alembic==1.10.4
-anyio==3.6.2
-argon2-cffi==21.3.0
-argon2-cffi-bindings==21.2.0
-arrow==1.2.3
-astroid==2.15.5
-asttokens==2.2.1
-attrs==23.1.0
-Babel==2.12.1
-backcall==0.2.0
-beautifulsoup4==4.12.2
-bleach==6.0.0
-certifi==2023.5.7
-cffi==1.15.1
-charset-normalizer==3.1.0
-click==8.1.3
-cloudpickle==2.2.1
-comm==0.1.3
-cycler==0.11.0
-databricks-cli==0.17.6
-debugpy==1.6.7
-decorator==5.1.1
-defusedxml==0.7.1
-dill==0.3.6
-dj-settings==4.2.1
-docker==6.0.1
-entrypoints==0.4
-executing==1.2.0
-fastjsonschema==2.16.3
-Flask==2.2.3
-fqdn==1.5.1
-future==0.18.3
-gitdb==4.0.10
-GitPython==3.1.31
-greenlet==2.0.2
-gunicorn==20.1.0
-hyperopt==0.2.7
-idna==3.4
-importlib-metadata==6.6.0
-ipykernel==6.5.0
-ipython==7.34.0
-ipython-genutils==0.2.0
-ipywidgets==8.0.6
-isoduration==20.11.0
-isort==5.12.0
-itsdangerous==2.1.2
-jedi==0.18.2
-Jinja2==3.1.2
-joblib==1.1.1
-json5==0.9.11
-jsonpointer==2.3
-jsonschema==4.17.3
-jupyter==1.0.0
-jupyter-console==6.4.4
-jupyter-events==0.6.3
-jupyter-ydoc==0.2.4
-jupyter_client==7.4.9
-jupyter_core==5.3.0
-jupyter_server==2.5.0
-jupyter_server_fileid==0.9.0
-jupyter_server_terminals==0.4.4
-jupyter_server_ydoc==0.8.0
-jupyterlab==3.6.3
-jupyterlab-pygments==0.2.2
-jupyterlab-widgets==3.0.7
-jupyterlab_server==2.22.1
-kiwisolver==1.4.4
-lazy-object-proxy==1.9.0
-MarkupSafe==2.1.2
-matplotlib==3.4.3
-matplotlib-inline==0.1.6
-mccabe==0.7.0
-mistune==2.0.5
-mlflow==1.26.1
-nbclassic==0.5.5
-nbclient==0.7.3
-nbconvert==7.3.1
-nbformat==5.8.0
-nest-asyncio==1.5.6
-networkx==3.1
-notebook==6.5.4
-notebook_shim==0.2.2
-numpy==1.25.0
-oauthlib==3.2.2
-packaging==23.1
-pandas==1.3.4
-pandocfilters==1.5.0
-parso==0.8.3
-pickleshare==0.7.5
-Pillow==9.5.0
-platformdirs==3.6.0
-prometheus-client==0.16.0
-prometheus-flask-exporter==0.22.3
-prompt-toolkit==3.0.38
-protobuf==3.13.0
-psutil==5.9.5
-pure-eval==0.2.2
-pvenv==1.0.0
-py4j==0.10.9.7
-pyarrow==8.0.0
-pycparser==2.21
-Pygments==2.15.1
-PyJWT==2.7.0
-pylint==2.17.4
-pyparsing==3.0.9
-pyrsistent==0.19.3
-python-dateutil==2.8.2
-python-json-logger==2.0.7
-pytz==2023.3
-PyYAML==6.0
-pyzmq==25.0.2
-qtconsole==5.4.2
-QtPy==2.3.1
-querystring-parser==1.2.4
-requests==2.30.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-scikit-learn==1.2.2
-scipy==1.10.1
-seaborn==0.12.2
-Send2Trash==1.8.0
-six==1.16.0
-smmap==5.0.0
-sniffio==1.3.0
-soupsieve==2.4.1
-SQLAlchemy==2.0.13
-sqlparse==0.4.4
-stack-data==0.6.2
-tabulate==0.9.0
-terminado==0.17.1
-threadpoolctl==2.0.0
-tinycss2==1.2.1
-tomli==2.0.1
-tomlkit==0.11.8
-tornado==6.3.1
-tqdm==4.65.0
-traitlets==5.9.0
-typing==3.7.4.3
-typing_extensions==4.5.0
-uri-template==1.2.0
-wcwidth==0.2.6
-webcolors==1.13
-webencodings==0.5.1
-websocket-client==1.5.1
-widgetsnbextension==4.0.7
-wrapt==1.15.0
-xgboost==1.7.1
-y-py==0.5.9
-ypy-websocket==0.8.2
-zipp==3.15.0
